refactored functions to seperate modules
